refactor: reorganize nar and narinfo helper logic into dedicated packages (#821)
Move NAR and NarInfo helper functions (file path generation and hash
validation) from the generic pkg/helper package to specialized pkg/nar
and pkg/narinfo packages. This change improves the internal architecture
by co-locating logic with the data types they operate on, making the
codebase more discoverable and maintainable.

Key changes:
- Created pkg/nar package with FilePath and ValidateHash logic.
- Created pkg/narinfo package with FilePath and ValidateHash logic.
- Updated pkg/storage/local and pkg/storage/s3 to use these new packages.
- Simplified pkg/helper to only contain the generic FilePathWithSharding logic.
- Migrated all relevant tests and updated import paths across the project.

(cherry picked from commit f1a147977549bc5e9e2d264a4e6a6bbe3dd52586)
