+++ pkg/helper/filepath.go
@@ -4,59 +4,10 @@ import (
 	"errors"
 	"fmt"
 	"path/filepath"
-	"regexp"
 )
 
-var (
-	// ErrInvalidHash is returned if the hash is invalid.
-	ErrInvalidHash = errors.New("invalid hash")
-
-	// hashRegexp is used to validate hashes.
-	hashRegexp = regexp.MustCompile(`^[a-z0-9]+$`)
-
-	// ErrInputTooShort is returned if the input is too short.
-	ErrInputTooShort = errors.New("input is less than 3 characters long")
-)
-
-// ValidateHash validates the given hash.
-func ValidateHash(hash string) error {
-	if !hashRegexp.MatchString(hash) {
-		return ErrInvalidHash
-	}
-
-	return nil
-}
-
-// NarInfoFilePath returns the path of the narinfo file given a hash.
-func NarInfoFilePath(hash string) (string, error) {
-	if len(hash) < 3 {
-		return "", fmt.Errorf("hash=%q: %w", hash, ErrInputTooShort)
-	}
-
-	if err := ValidateHash(hash); err != nil {
-		return "", err
-	}
-
-	return FilePathWithSharding(hash + ".narinfo")
-}
-
-// NarFilePath returns the path of the nar file given a hash and an optional compression.
-func NarFilePath(hash, compression string) (string, error) {
-	if len(hash) < 3 {
-		return "", fmt.Errorf("hash=%q: %w", hash, ErrInputTooShort)
-	}
-
-	if err := ValidateHash(hash); err != nil {
-		return "", err
-	}
-
-	fn := hash + ".nar"
-	if compression != "" {
-		fn += "." + compression
-	}
-
-	return FilePathWithSharding(fn)
-}
+// ErrInputTooShort is returned if the input is less than 3 characters long.
+var ErrInputTooShort = errors.New("input is less than 3 characters long")
 
 // FilePathWithSharding returns the path to a file with sharding.
 func FilePathWithSharding(fn string) (string, error) {
