+++ pkg/helper/filepath_test.go
@@ -11,82 +11,6 @@ import (
 	"github.com/kalbasit/ncps/pkg/helper"
 )
 
-func TestNarInfoFilePath(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		hash string
-		path string
-	}{
-		{hash: "abc123", path: filepath.Join("a", "ab", "abc123.narinfo")},
-	}
-
-	for _, test := range []string{"", "a", "ab"} {
-		t.Run(fmt.Sprintf("NarInfoFilePath(%q) should return error", test), func(t *testing.T) {
-			t.Parallel()
-
-			_, err := helper.NarInfoFilePath(test)
-			assert.ErrorContains(t, err, "is less than 3 characters long")
-		})
-	}
-
-	for _, test := range tests {
-		t.Run(fmt.Sprintf("NarInfoFilePath(%q) -> %q", test.hash, test.path), func(t *testing.T) {
-			t.Parallel()
-
-			path, err := helper.NarInfoFilePath(test.hash)
-			require.NoError(t, err)
-			assert.Equal(t, test.path, path)
-		})
-	}
-
-	t.Run("NarInfoFilePath with invalid hash", func(t *testing.T) {
-		t.Parallel()
-
-		_, err := helper.NarInfoFilePath("abc!@#")
-		assert.ErrorIs(t, err, helper.ErrInvalidHash)
-	})
-}
-
-func TestNarFilePath(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		hash        string
-		compression string
-		path        string
-	}{
-		{hash: "abc123", compression: "", path: filepath.Join("a", "ab", "abc123.nar")},
-		{hash: "def456", compression: "xz", path: filepath.Join("d", "de", "def456.nar.xz")},
-	}
-
-	for _, test := range []string{"", "a", "ab"} {
-		t.Run(fmt.Sprintf("NarFilePath(%q) should return error", test), func(t *testing.T) {
-			t.Parallel()
-
-			_, err := helper.NarFilePath(test, "")
-			assert.ErrorContains(t, err, "is less than 3 characters long")
-		})
-	}
-
-	for _, test := range tests {
-		t.Run(fmt.Sprintf("NarFilePath(%q, %q) -> %q", test.hash, test.compression, test.path), func(t *testing.T) {
-			t.Parallel()
-
-			path, err := helper.NarFilePath(test.hash, test.compression)
-			require.NoError(t, err)
-			assert.Equal(t, test.path, path)
-		})
-	}
-
-	t.Run("NarFilePath with invalid hash", func(t *testing.T) {
-		t.Parallel()
-
-		_, err := helper.NarFilePath("abc!@#", "")
-		assert.ErrorIs(t, err, helper.ErrInvalidHash)
-	})
-}
-
 func TestFilePathWithSharding(t *testing.T) {
 	t.Parallel()
